Reworked main loop logic as an FSM. This folds the "send" and the "cull" loops together.
My rationalle here is that handling interrupts (whilst getting the interval timing correct) is difficult, and I think it best to use the same structure for both situations, rather than duplicating that elsewhere.

Previously the culled timeouts were postponed until an interrupt, which was incorrect behaviour.

I believe the timing is incorrect here (the intervals seem wildly off), but at least it's all in one place.